test_deploy/requirements.txt
functions-framework==3.5.0
google-cloud-storage==2.10.0
google-cloud-bigquery==3.13.0

